Move AfkFrequencyComponent file inside R-TypeLogicLibrary/Server #227
