+++ libs/R-TypeLogic/Global/Components/CMakeLists.txt
@@ -34,6 +34,5 @@ set(SRC_RTYPE_LOGIC_COMPONENTS
     ${SRC_RTYPE_LOGIC_COMPONENTS_ROOT}/NewlyCreated.hpp
     ${SRC_RTYPE_LOGIC_COMPONENTS_ROOT}/LayerLvL.hpp
     ${SRC_RTYPE_LOGIC_COMPONENTS_ROOT}/DisconnectableComponent.hpp
-    ${SRC_RTYPE_LOGIC_COMPONENTS_ROOT}/AfkFrequencyComponent.hpp
     PARENT_SCOPE
 )
