+++ libs/R-TypeLogic/Global/Systems/UpdateClockSystem.cpp
@@ -8,7 +8,7 @@
 #include "UpdateClockSystem.hpp"
 #include <chrono>
 #include "R-TypeLogic/Global/Components/ShootingFrequencyComponent.hpp"
-#include "R-TypeLogic/Global/Components/AfkFrequencyComponent.hpp"
+#include "R-TypeLogic/Server/Components/AfkFrequencyComponent.hpp"
 #include "R-TypeLogic/Global/SharedResources/GameClock.hpp"
 
 using namespace ecs;
