+++ libs/R-TypeLogic/Server/Components/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SRC_RTYPE_LOGIC_COMPONENTS_ROOT ${PROJECT_SOURCE_DIR}/libs/R-TypeLogic/Server/Components)
+
+set(SRC_RTYPE_LOGIC_SERVER_COMPONENTS
+    ${SRC_RTYPE_LOGIC_COMPONENTS_ROOT}/AfkFrequencyComponent.hpp
+    PARENT_SCOPE
+)
